doc: Add User Guide
Oddvar Lia created the MindMap and exported it.
I (Sindre Nistad) merely uploaded / added it to git.

The User Guide is created, and maintained using [MindManager](https://www.mindmanager.com/en/) and is then exported as HTML to `index.html`.

GitHub Pages does not deal well with symbolic links at all.
Originally, `index.html` was named `APS_user_guide.html` and was symlinked to `index.html`, but that caused the deployment to fail.

Co-authored-by: Oddvar Lia <olia@equinor.com>