Added optional keyword to add noise to extrapolated values.
The purpose is to avoid that all realizations in ERT will get 0 standard deviation
in grid cells that only exist in ERTBOX but not in geogrid. 0 standard deviation
in ERT prior ensemble can per today cause errors and warnings from numpy
operations related to parameter update in particular for adaptive localisation.